+++ concourse/models.go
@@ -83,7 +83,7 @@ type SecretValues map[string]secretValue // key is secret "<mount>-<path>", and 
 type secretValue map[string]any // key-value pairs would be arbitrary for kv1 and kv2, but are standardized schema for credential generators
 
 // lease id validation regex for below constructor
-var leaseIDRegex = regexp.MustCompile(`\w{8}-\w{4}-\w{4}-\w{4}-\w{12}`)
+var leaseIDRegex = regexp.MustCompile(`[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$`)
 
 // checkRequest constructor with pipeline param as io.Reader but typically os.Stdin *os.File input because concourse
 func NewCheckRequest(pipelineJSON io.Reader) (*checkRequest, error) {